First stab at getting integ tests running in GH actions

+++ local-dev/bootstrap-local-kind-cluster.sh
@@ -10,17 +10,21 @@ export AWS_ACCOUNT_ID=`aws sts get-caller-identity --query 'Account' --output te
 
 # Create required AWS resources
 
-echo Deploying CloudFormation stack and setting git credentials
+if [[ ! -v CI ]]; then
+    echo Deploying CloudFormation stack with prerequisite resources
 
-aws cloudformation deploy --stack-name flux-cfn-controller-resources --region $AWS_REGION --template-file examples/resources.yaml --capabilities CAPABILITY_NAMED_IAM
+    aws cloudformation deploy --stack-name flux-cfn-controller-resources --region $AWS_REGION --template-file examples/resources.yaml --capabilities CAPABILITY_NAMED_IAM
 
-existing_creds=`aws iam list-service-specific-credentials --user-name flux-git --service-name codecommit.amazonaws.com --query 'ServiceSpecificCredentials'`
-empty_creds="[]"
+    existing_creds=`aws iam list-service-specific-credentials --user-name flux-git --service-name codecommit.amazonaws.com --query 'ServiceSpecificCredentials'`
+    empty_creds="[]"
 
-if [ "$existing_creds" = "$empty_creds" ]; then
-    new_creds=`aws iam create-service-specific-credential --user-name flux-git --service-name codecommit.amazonaws.com --query 'ServiceSpecificCredential' --output json`
-    aws secretsmanager put-secret-value --region $AWS_REGION --secret-string "$new_creds" --secret-id flux-git-credentials
+    if [ "$existing_creds" = "$empty_creds" ]; then
+        new_creds=`aws iam create-service-specific-credential --user-name flux-git --service-name codecommit.amazonaws.com --query 'ServiceSpecificCredential' --output json`
+        aws secretsmanager put-secret-value --region $AWS_REGION --secret-string "$new_creds" --secret-id flux-git-credentials
+    fi
 fi
+
+echo Setting up git repository for CloudFormation templates
 
 creds=`aws secretsmanager get-secret-value --region $AWS_REGION --secret-id flux-git-credentials --query 'SecretString' --output text`
 
@@ -71,18 +75,20 @@ flux create secret git cfn-template-repo-auth \
     --username=$CODECOMMIT_USERNAME \
     --password=$CODECOMMIT_PASSWORD
 
-rm -rf patch-local-cluster
-mkdir patch-local-cluster
-cd patch-local-cluster
-git clone https://$CODECOMMIT_USERNAME:$CODECOMMIT_PASSWORD@git-codecommit.$AWS_REGION.amazonaws.com/v1/repos/my-flux-configuration
-cd my-flux-configuration
-git apply ../../local-dev/local-flux-dev-config.patch
-git add flux-system
-git commit -m "Expose source controller locally"
-git push
-cd ../..
-rm -rf patch-local-cluster
-flux reconcile source git flux-system
+if [[ ! -v CI ]]; then
+    rm -rf patch-local-cluster
+    mkdir patch-local-cluster
+    cd patch-local-cluster
+    git clone https://$CODECOMMIT_USERNAME:$CODECOMMIT_PASSWORD@git-codecommit.$AWS_REGION.amazonaws.com/v1/repos/my-flux-configuration
+    cd my-flux-configuration
+    git apply ../../local-dev/local-flux-dev-config.patch
+    git add flux-system
+    git commit -m "Expose source controller locally"
+    git push
+    cd ../..
+    rm -rf patch-local-cluster
+    flux reconcile source git flux-system
+fi
 
 # Install CFN controller types
 
@@ -102,10 +108,7 @@ flux get all
 
 echo Installing credentials into the kind cluster
 
-# Note that this ECR token will expire, so this is only for development use
-#kubectl delete secret docker-registry ecr-cred -n flux-system --ignore-not-found
-#kubectl create secret docker-registry ecr-cred --docker-server=$AWS_ACCOUNT_ID.dkr.ecr.$AWS_REGION.amazonaws.com --docker-username=AWS --docker-password=`aws ecr get-login-password --region $AWS_REGION` -n flux-system
-#kubectl patch serviceaccount default -p "{\"imagePullSecrets\": [{\"name\": \"ecr-cred\"}]}" -n flux-system
-
 kubectl delete secret aws-creds -n flux-system --ignore-not-found
-kubectl create secret generic aws-creds -n flux-system --from-file ~/.aws/credentials
+if [[ ! -v CI ]]; then
+    kubectl create secret generic aws-creds -n flux-system --from-file ~/.aws/credentials
+fi
